+++ app/web/sample/src/services.ts
@@ -1,6 +1,6 @@
 import { useState, useEffect } from "react"
 
-import { configure } from "~/packages/atom.io/src"
+import { atom, selector } from "~/packages/atom.io/src"
 import { composeStoreHooks } from "~/packages/atom.io/src/react"
 
 const { useStore, useO, useI, useIO } = composeStoreHooks({
@@ -8,8 +8,27 @@ const { useStore, useO, useI, useIO } = composeStoreHooks({
   useEffect,
 })
 
-configure({
-  logger: console,
+export { useStore, useO, useIO, useI }
+
+export const dividendState = atom<number>({
+  key: `dividend`,
+  default: 1,
 })
 
-export { useStore, useO, useIO, useI }
+export const divisorState = atom<number>({
+  key: `divisor`,
+  default: 2,
+})
+
+export const quotientState = selector<number>({
+  key: `quotient`,
+  get: ({ get }) => {
+    const divisor = get(divisorState)
+    const dividend = get(dividendState)
+    return dividend / divisor
+  },
+  set: ({ get, set }, newValue) => {
+    const divisor = get(divisorState)
+    set(dividendState, newValue * divisor)
+  },
+})
